move search base selection to find dialog
\not needed in filter widget
