+++ src/admc/filter_widget/filter_widget.cpp
@@ -25,31 +25,13 @@
 #include "utils.h"
 
 #include <QLabel>
-#include <QComboBox>
 #include <QTabWidget>
 #include <QGridLayout>
 #include <QPushButton>
 
-// TODO: search base input should be in parent widget because its not always needed. For example it's not needed when filter widget will be used to filter contents widget.
-
 FilterWidget::FilterWidget()
 : QWidget()
 {
-    auto search_base_combo_label = new QLabel(tr("In:"));
-    search_base_combo = new QComboBox();
-
-    // TODO: technically, entire directory does NOT equal to the domain. In cases where we're browsing multiple domains at the same time (or maybe some other situations as well), we'd need "Entire directory" AND all of domains. Currently search base is set to domain anyway, so would need to start from reworking that.
-    search_base_combo->addItem(tr("Entire directory"), AD()->search_base());
-
-    const QString users_dn = "CN=Users," + AD()->search_base();
-    search_base_combo->addItem("Users", users_dn);
-
-    auto custom_search_base_button = new QPushButton(tr("Browse"));
-    connect(
-        custom_search_base_button, &QAbstractButton::clicked,
-        this, &FilterWidget::on_custom_search_base);
-
-
     tab_widget = new QTabWidget();
 
     auto normal_tab = new FilterWidgetNormalTab();
@@ -63,11 +45,6 @@ FilterWidget::FilterWidget()
     auto layout = new QGridLayout();
     setLayout(layout);
 
-    const int search_base_combo_row = layout->rowCount();
-    layout->addWidget(search_base_combo_label, search_base_combo_row, 0);
-    layout->addWidget(search_base_combo, search_base_combo_row, 1);
-    layout->addWidget(custom_search_base_button, search_base_combo_row, 2);
-
     const int tab_widget_row = layout->rowCount();
     const int tab_widget_col_span = layout->columnCount();
     layout->addWidget(tab_widget, tab_widget_row, 0, 1, tab_widget_col_span);
@@ -76,29 +53,6 @@ FilterWidget::FilterWidget()
 QString FilterWidget::get_filter() const {
     const FilterWidgetTab *current_tab = get_current_tab();
     return current_tab->get_filter();
-}
-
-QString FilterWidget::get_search_base() const {
-    const int index = search_base_combo->currentIndex();
-    const QVariant item_data = search_base_combo->itemData(index);
-
-    return item_data.toString();
-}
-
-void FilterWidget::on_custom_search_base() {
-    // TODO: maybe need some other classes?
-    const QList<QString> selecteds = SelectDialog::open({CLASS_CONTAINER, CLASS_OU});
-
-    if (!selecteds.isEmpty()) {
-        const QString selected = selecteds[0];
-        const QString name = dn_get_rdn(selected);
-
-        search_base_combo->addItem(name, selected);
-
-        // Select newly added search base
-        const int new_base_index = search_base_combo->count() - 1;
-        search_base_combo->setCurrentIndex(new_base_index);
-    }
 }
 
 const FilterWidgetTab *FilterWidget::get_current_tab() const {
